move makeup-specific code to the highlighter module
