+++ lib/ex_doc/highlighter.ex
@@ -1,6 +1,26 @@
 defmodule ExDoc.Highlighter do
   @moduledoc false
 
+  alias Makeup.Formatters.HTML.HTMLFormatter
+
+  @external_resource "assets/dist/ex_doc_makeup.css"
+
+  @assets [
+    # Read the CSS from the included file.
+    # This allows us to have a custom CSS theme not included in Makeup
+    # that supports both "day mode" and "night mode".
+    {"dist/ex_doc_makeup-css.css", File.read!("assets/dist/ex_doc_makeup.css")},
+    # Get the Javascript snippet directly from Makeup.
+    # If there is any need to customize it further, we can add a "ex_doc_makeup.js" file.
+    {"dist/ex_doc_makeup-js.js", HTMLFormatter.group_highlighter_javascript()}
+  ]
+
+  def assets(_), do: @assets
+
+  def before_closing_head_tag(_), do: ~S(<link rel="stylesheet" href="dist/ex_doc_makeup-css.css"/>)
+
+  def before_closing_body_tag(_), do: ~S(<script src="dist/ex_doc_makeup-js.js"></script>)
+
   # If new lexers are available, add them here:
   defp pick_language_and_lexer(""), do: {"elixir", Makeup.Lexers.ElixirLexer}
   defp pick_language_and_lexer("elixir"), do: {"elixir", Makeup.Lexers.ElixirLexer}
