+++ lib/ex_doc/markdown/earmark.ex
@@ -2,30 +2,14 @@ defmodule ExDoc.Markdown.Earmark do
   @moduledoc """
   ExDoc extension for the Earmark MarkDown parser.
   """
-
-  alias Makeup.Formatters.HTML.HTMLFormatter
-
   @behaviour ExDoc.Markdown
 
-  @external_resource "assets/dist/ex_doc_makeup.css"
-
   # Callback implementations
+  def assets(arg), do: ExDoc.Highlighter.assets(arg)
 
-  @assets [
-    # Read the CSS from the included file.
-    # This allows us to have a custom CSS theme not included in Makeup
-    # that supports both "day mode" and "night mode".
-    {"dist/ex_doc_makeup-css.css", File.read!("assets/dist/ex_doc_makeup.css")},
-    # Get the Javascript snippet directly from Makeup.
-    # If there is any need to customize it further, we can add a "ex_doc_makeup.js" file.
-    {"dist/ex_doc_makeup-js.js", HTMLFormatter.group_highlighter_javascript()}
-  ]
+  def before_closing_head_tag(arg), do: ExDoc.Highlighter.before_closing_head_tag(arg)
 
-  def assets(_), do: @assets
-
-  def before_closing_head_tag(_), do: ~S(<link rel="stylesheet" href="dist/ex_doc_makeup-css.css"/>)
-
-  def before_closing_body_tag(_), do: ~S(<script src="dist/ex_doc_makeup-js.js"></script>)
+  def before_closing_body_tag(arg), do: ExDoc.Highlighter.before_closing_body_tag(arg)
 
   def configure(_), do: :ok
