+++ test/ex_doc/formatter/html/templates_test.exs
@@ -227,8 +227,6 @@ defmodule ExDoc.Formatter.HTML.TemplatesTest do
   test "module_page outputs the functions and docstrings" do
     content = get_module_page([CompiledWithDocs])
 
-    IO.puts(content)
-
     # Title and headers
     assert content =~ ~r{<title>CompiledWithDocs [^<]*</title>}
     assert content =~ ~r{<h1>\s*<small class="visible-xs">Elixir v1.0.1</small>\s*CompiledWithDocs\s*}
